moved some web test from legacy itv to itvx
